refactor(grades): one grade system owned by classification
Wave 4B of the 2026-08-22 review implementation plan:

- V9GradeSchema/V9Grade alias the canonical report-card grade enum
  instead of restating it; the unclamped scoreToV9Grade and forked
  v9GradeRange collapse into the clamping scoreToGrade/gradeRange.
- Grade range metadata, Tailwind palettes, and radar colors move into
  shared/lib/classification/grades.ts (per the classification
  ownership rule); report-card-core, report-card-ui, and the
  safety-score map generator consume the canonical exports.
- The report-cards.ts and report-card-resilience.ts compatibility
  shims are deleted with all importers migrated; lockstep copy tests
  are replaced by alias-identity assertions.
- Regenerated the Safety Score V9 evaluation-build manifest for the
  relocated grade modules.
