+++ shared/lib/__tests__/report-card-core.test.ts
@@ -3,11 +3,10 @@ import {
   getReportCardGradeRank,
   GRADE_THRESHOLDS,
   REPORT_CARD_GRADE_RANK,
-  scoreToV9Grade,
   scoreToGrade,
   UNKNOWN_REPORT_CARD_GRADE_RANK,
   V9_GRADE_THRESHOLDS,
-} from "../report-cards";
+} from "../report-card-core";
 import { V9_CANDIDATE_POLICY_V1 } from "../safety-score-v9/policy";
 
 describe("scoreToGrade", () => {
@@ -20,12 +19,12 @@ describe("scoreToGrade", () => {
     expect(V9_GRADE_THRESHOLDS).toEqual(expected);
   });
 
-  it("keeps the V9 grade conversion bound to policy thresholds", () => {
+  it("keeps the grade conversion bound to policy thresholds", () => {
     for (const { grade, minScore } of V9_CANDIDATE_POLICY_V1.policy.semantic.formula.gradeThresholds) {
-      expect(scoreToV9Grade(minScore)).toBe(grade);
-      if (minScore > 0) expect(scoreToV9Grade(minScore - 0.1)).not.toBe(grade);
+      expect(scoreToGrade(minScore)).toBe(grade);
+      if (minScore > 0) expect(scoreToGrade(minScore - 0.1)).not.toBe(grade);
     }
-    expect(scoreToV9Grade(null)).toBe("NR");
+    expect(scoreToGrade(null)).toBe("NR");
   });
 
   it("returns NR for null", () => {
