+++ shared/lib/report-card-policy.ts
@@ -3,9 +3,11 @@ import type {
   CollateralQuality,
   CustodyModel,
   GovernanceType,
+  ReserveSlice,
   ReserveRisk,
 } from "../types";
 import type { StablecoinClientMeta } from "../types/stablecoin-client-meta";
+import { roundScore } from "./math";
 
 type ResilienceDefaults = {
   collateralQuality: CollateralQuality;
@@ -58,6 +60,13 @@ export const RESERVE_QUALITY_SCORE: Record<ReserveRisk, number> = {
   high: 25,
   "very-high": 5,
 };
+
+export function computeCollateralQualityFromReserves(reserves: ReserveSlice[]): number {
+  const totalPct = reserves.reduce((sum, reserve) => sum + reserve.pct, 0);
+  if (totalPct === 0) return 0;
+  const weighted = reserves.reduce((sum, reserve) => sum + reserve.pct * (RESERVE_QUALITY_SCORE[reserve.risk] ?? 0), 0);
+  return roundScore(weighted / totalPct);
+}
 
 export function inferResilienceDefaults(backing: BackingType, governance: GovernanceType): ResilienceDefaults {
   return DEFAULT_RESILIENCE_FACTORS[`${backing}:${governance}`];
