+++ worker/src/lib/collateral-drift.ts
@@ -1,6 +1,6 @@
 import { isReserveDriftThresholdExceeded } from "@shared/lib/status-thresholds";
 import { ACTIVE_STABLECOINS } from "@shared/lib/stablecoins/registry";
-import { computeCollateralQualityFromReserves } from "@shared/lib/report-cards";
+import { computeCollateralQualityFromReserves } from "@shared/lib/report-card-policy";
 import type { ReserveSlice, StablecoinMeta } from "@shared/types/core";
 import { loadFreshIndependentLiveReserveMap } from "./live-reserves-store";
